Reported bug fix.
Our automated buildd log filter[1] detected a problem that will cause                                   
your package to segfault on architectures where the size of a pointer                                   
is greater than the size of an integer, such as ia64 and amd64.                                         
                                                                                                        
This is often due to a missing function prototype definition.                                           
                                                                                                        
[1]http://people.debian.org/~dannf/check-implicit-pointer-functions                                     
                                                                                                        
Function `cgdb_malloc' implicitly converted to pointer at kui_driver.c:122


git-svn-id: https://cgdb.svn.sourceforge.net/svnroot/cgdb/cgdb/trunk@878 61d65b3c-fd13-0410-b770-9cb5f9339520