Clean up directory structure.
Previously split src/ and include/ directories in all components, which
is not really a helpful way of doing things.  Moved a bunch of libraries
to lib/ instead of letting them live all over the place.  The primary
top-level source directories now are lib/ and cgdb/.

Had to add header files to the Makefile.am files so that they would be
distributed properly.