Fix issue #40 - map commands are improved.
This is honestly just a slight improvement over what was previously
existing. Now a command like,
  imap a b
in the ~/.cgdb/cgdbrc file won't map a to 'b<nl>' it will just map
it to 'b'. However, there are probably lots of other characters
that the map command should ignore. I haven't resolved this at this
point in time. I only handled, ' ', \t, \f, \v, \n and \r.