Colorize color names in :help file
Skip the word Black since it won't show up on a black background.

Bob: I moved the check for DEBUGWINCOLOR from highlight_groups.cpp
to scroller.cpp. That's because when color was true, and debugwincolor was
false, the :help syntax highlighter would not show colored names due
to the fact that ansi_color was set to false (which was due to the fact that
ansi_color was dependent on DEBUGWINCOLOR).

Signed-off-by: Michael Sartain <mikesart@gmail.com>
Signed-off-by: Bob Rossi <bob@brasko.net>