Use wnoutrefresh/doupdate instead of wrefresh when drawing ncurses windows
From the ncurses doupdate() man page:

> The routine wrefresh works by first calling wnoutrefresh, which copies
the named window to the virtual screen, and then calling doupdate, which
compares the virtual screen to the physical screen and does the actual
update. If the programmer wishes to output several windows at once, a
series of calls to wrefresh results in alternating calls to wnoutrefresh
and doupdate, causing several bursts of output to the screen. By first
calling wnoutrefresh for each window, it is then possible to call
doupdate once, resulting in only one burst of output, with fewer total
characters transmitted and less CPU time used.

This eliminates much of the flashing when scrolling through large
amounts of output.

Signed-off-by: Michael Sartain <mikesart@gmail.com>
Signed-off-by: Bob Rossi <bob@brasko.net>