Simplify kui internal implementation.
Previously, a kuictx instance had a list of map sets. That is, a single
kui context could take as input many map sets and return the appropriate
keys based on that. The member looked like,
-    std_list_ptr kui_map_set_list;

The kui context has been changed to only support a single map set.
+    struct kui_map_set *map_set;

CGDB never used multiple map sets for a single kui context and the
code was never tested with multiple map sets. Since the functionality
was never used or tested, I'm removing it, and simplifying the
implementation.

Signed-off-by: Bob Rossi <bob@brasko.net>