Update ability to display currently selected or executing line.
Added support for 'executinglinedisplay' and 'selectedlinedisplay'
configuration options. See documentation for full explanation.
This allows you to configure how CGDB displays both the currently
executing line and the currently selected line.
The default for executing line is set to longarrow.
The default for selected line is set to block.

The 'arrowstyle' option has been deprecated (but still supported)
and users should use the 'executinglinedisplay' option instead.

Signed-off-by: Bob Rossi <bob@brasko.net>