Refactor sources.cpp:source_display (add ExecutingLineNr hl group)
Previously, the source display function would try to print lines based on
the line kind: selected line, executing line, breakpoint line, other.

I've updated the code to print the line based on the section the line
is on. The sections are # | marker text. This is documented in the
source_display function in the implementation.

I got inspiration from Michael Sartain in a patch that he applied to
his local fork. It no longer cleanly applied here, so I took some
of that work, mainly the changes to
    cgdb/filedlg.cpp
    cgdb/highlight.cpp
    cgdb/highlight.h

Also added ExecutingLineNr, which highlights the currently executing
line.

Signed-off-by: Bob Rossi <bob@brasko.net>