Fix a memory corruption bug in scroller prompt.
I'm not exactly sure how, but the realloc was not called properly.
rv would be fine going into the call and after the call there would
occasionally be some garbage characters in the output buffer.

Either way, it didn't matter. Now that CGDB is C++, I allocate the
buffer on the stack and initialize it to zero. At the end of the call
I malloc only the memory necessary.

I also removed the functionality to trim whitespace at the end of
the scroller line. I'm not exactly sure how this code was useful.
I think it might have been a previous attempt to solve the prompt
problem I solved in the last few commits.

Signed-off-by: Bob Rossi <bob@brasko.net>