Fix regression in gdb window display when running debugger commands.
In cgdb-0.6.8 if you hit F8 (to issue a next) or type :next,
a new line would be put into the gdb console window. In master/
that was not occurring. This means that if you hit F8 over and over,
you would end up with something like this in the gdb window,
  (gdb) (gdb) (gdb) etc.

To the best of my knowledge this patch makes master/ behave like
cgdb-0.6.8.

Signed-off-by: Bob Rossi <bob@brasko.net>