Fix unnecessary flashing while redrawing.
I was trying to work around a bug (I think) in xfce4-terminal.
You can see the issue here,
  http://lists.gnu.org/archive/html/bug-ncurses/2016-08/msg00001.html

The code that I added to work around that bug made cgdb flash when
holding down F8, to do next after next. It also makes it flash in
other conditions when the gdb window has moved between actions.

Removing the work around. Not sure how to make cgdb work great
regarding the cursor with xfce4-terminal, but it works properly
with gnome-terminal and xterm.

Signed-off-by: Bob Rossi <bob@brasko.net>