Fix scroller display issue with long lines.
If the scroller had long lines, CGDB may have been showing to little
text in the scroller. This patch is intended to resolve that issue.

Basically, the scroller keeps track of how many lines of text
it should show to support the ctrl-l feature. Some lines take up
more than one row in the scroller. This patch allows the scroller
to print the appropriate number of rows based on the number of
lines that should be printed. Previously, we assumed the number
of lines was the same as the number of rows.

Signed-off-by: Bob Rossi <bob@brasko.net>