Rewrite the annotations module.
I tried to do this a little at a time, but the commands, tgdb, a2-tgdb
and state-machine files were tightly coupled. The new annotations package
is very independent and much simpler to read.

I tested to the best of my ability that each annotation, when hit,
still behaves properly.

I removed the TGDB_INFERIOR_EXITED response because it currently
was not being used by cgdb. That was probably a bug. However, when
we need it, we can add it back in.

Signed-off-by: Bob Rossi <bob@brasko.net>