Start gdb with -ex <command> instead of -x <filename>
Currently, the gdb init file only had in it,
    set annotate=2
    set height=0

So I pass those with -ex instead. It's simpler, and doesn't involve
writing a file and passing a filename to gdb.

I'm investigating if it's also possible to remove the set annotate=2
line, since we already pass --annotate=2 on the command line.

Signed-off-by: Bob Rossi <bob@brasko.net>