Simplify tgdb_process further.
Previously, tgdb would send the quit command on EOF from GDB or
on SIGCHLD of GDB from OS. The SIGCHLD would have the child return
status and the EOF would not. I think the order which you could get
these is sporadic and I'm not honestly sure which one I should honor.

I've decided to continue to cleanup the SIGCHLD processes but only
quit tgdb on EOF from GDB. I don't feel strongly about this and
would be happy to change it if the need arises.

CGDB is currently not using the fields from the TGDB_QUIT command,
so I simply removed them from the structure.

Signed-off-by: Bob Rossi <bob@brasko.net>