Add the annotations parser_at_prompt callback.
This allows events to occur when the parser prompt is hit.

One need for this was to set the flag in tgdb determining if a command
has been completed. Simply querying if the annotations parser is at that
state wasn't correct, because it would return it was, even if it never
moved away from that state.

Signed-off-by: Bob Rossi <bob@brasko.net>