Improve how CGDB reports to the user an error occurred.
Previously, if CGDB logged anything to .cgdb/logs/cgdb_log1.txt
then CGDB would report an error occurred to the user.
This meant that only errors could be written to the file.

Now, CGDB only reports an error occurred to the user if clog_error
is invoked. This means that warnings or info can now be put into
the file.

Signed-off-by: Bob Rossi <bob@brasko.net>