configure: fix error when setting LEX
In case LEX is set to some custom path when running `configure`, the
configuration would fail, since it requires $LEX to contain either
'flex' or 'lex'.

This change modifies the configure script to check the basename of the
value in $LEX, allowing to specify a path to the lexer in a custom
location.

Signed-off-by: Ton van den Heuvel <tonvandenheuvel@gmail.com>