Fix issue #260
breakpoint operations stop working after F5 or F6 or F7 or F8 or F12

Typically, when CGDB sends commands to GDB it will do it through the
new-ui mi channel. It then has to wait for gdb to respond before it can
send the next command.

However, debugger comamnds are instead sent to the gdb console window.
CGDB can send as many of these as it would like at once.

The bug was that CGDB was sending the debugger command to the console
window but expecting the new-ui mi channel to respond. This was never
going to happen, thus no new commands could be run through gdb.
The fix was to tell CGDB to not expect GDB to respond in this scenario.

Signed-off-by: Bob Rossi <bob@brasko.net>