Fix issue #303 - CGDB crashes while resizing window
Previously if a lot of text existed in the gdb window, then a resize
could trigger a crash. Now no crash will occur.

CGDB now simply resizes the vterm, and vterm does not support reflow,
so the text is not redrawn as expected when the columns grow or shrink.
This may be noticable when switching between horizontal and vertical mode.
There is a feature request for reflow support in vterm.

The decision was made to switch from deleting and recreating a vterm
instance with all previous text, to just resizing it, because recreating
the vterm with a large amount of text turned out to be very inefficient.
Resizing would take 1 second with only tens of thousands of lines of data.

Signed-off-by: Bob Rossi <bob@brasko.net>