Add ', to jump to the executing line across files
The '. jump command only resolves the currently executing line in the
currently displayed buffer. When navigating source files, it's sometimes
convenient to be able to jump back to the currently executing line
across files. I've implemented ', as a variation of '. to do that.

Alternatively, if preferred, I can also implement the feature in the
original '. command.

Signed-off-by: Pietro Cerutti <gahr@gahr.ch>