Allow toggling marks with the `m` cgdb command
This enhances the `m` command so that setting the same marker again at
the same location has the effect of removing the mark.

As a side note, a location can have many marks, and the mark displayed
is the one with the lowest letter. We might want to change that so that
the most recent mark is displayed, but this requires more fundamental
changes to how marks are managed. I think I'd do that as a separate PR.

Issue #319

Signed-off-by: Pietro Cerutti <pcerutti2@bloomberg.net>